+++ src/common/classes/am-options.class.js
@@ -20,6 +20,7 @@ export class amOptions extends AccountsCore.Options {
         editCloseAfterNew: true,
         editIdentTopTemplate: null,
         editTabsFn: null,
+        feedNow: true,
         listActiveCheckboxes: false,
         listFeedNow: true,
     };
@@ -59,6 +60,7 @@ export class amOptions extends AccountsCore.Options {
      */
     editCloseAfterNew( value ){
         logger.verbose({ verbosity: AccountsManager.configure().verbosity, against: AccountsManager.C.Verbose.FUNCTIONS }, 'amOptions.editCloseAfterNew()', arguments );
+        logger.warning( '\'editCloseAfterNew\' instanciation option is deprecated starting with v2.5 in favor of setupEditor({ \'closeAfterNew\' }). You should update your code.' );
         return this.base_gsBoolFn( 'editCloseAfterNew', value, { default: amOptions._defaults.editCloseAfterNew });
     }
 
@@ -69,6 +71,7 @@ export class amOptions extends AccountsCore.Options {
      */
     editIdentTopTemplate( value ){
         logger.verbose({ verbosity: AccountsManager.configure().verbosity, against: AccountsManager.C.Verbose.FUNCTIONS }, 'amOptions.editIdentTopTemplate()', arguments );
+        logger.warning( '\'editIdentTopTemplate\' instanciation option is deprecated starting with v2.5 in favor of setupEditor({ \'identTopTemplate\' }). You should update your code.' );
         return this.base_gsStringFn( 'editIdentTopTemplate', value, { default: amOptions._defaults.editIdentTopTemplate });
     }
 
@@ -79,6 +82,7 @@ export class amOptions extends AccountsCore.Options {
      */
     editTabsFn( value ){
         logger.verbose({ verbosity: AccountsManager.configure().verbosity, against: AccountsManager.C.Verbose.FUNCTIONS }, 'amOptions.editTabsFn()', arguments );
+        logger.warning( '\'editTabsFn\' instanciation option is deprecated starting with v2.5 in favor of setupEditor({ \'tabsFn\' }). You should update your code.' );
         return this.base_gsFn( 'editTabsFn', value, { default: amOptions._defaults.editTabsFn });
     }
 
@@ -89,6 +93,7 @@ export class amOptions extends AccountsCore.Options {
      */
     editWithGlobalRoles( value ){
         logger.verbose({ verbosity: AccountsManager.configure().verbosity, against: AccountsManager.C.Verbose.FUNCTIONS }, 'amOptions.editWithGlobalRoles()', arguments );
+        logger.warning( '\'editWithGlobalRoles\' instanciation option is deprecated starting with v2.5 in favor of setupEditor({ \'withGlobalRoles\' }). You should update your code.' );
         return this.base_gsBoolFn( 'editWithGlobalRoles', value, { default: amOptions._defaults.editWithGlobalRoles });
     }
 
@@ -99,7 +104,18 @@ export class amOptions extends AccountsCore.Options {
      */
     editWithScopedRoles( value ){
         logger.verbose({ verbosity: AccountsManager.configure().verbosity, against: AccountsManager.C.Verbose.FUNCTIONS }, 'amOptions.editWithScopedRoles()', arguments );
+        logger.warning( '\'editWithScopedRoles\' instanciation option is deprecated starting with v2.5 in favor of setupEditor({ \'withScopedRoles\' }). You should update your code.' );
         return this.base_gsBoolFn( 'editWithScopedRoles', value, { default: amOptions._defaults.editWithScopedRoles });
+    }
+
+    /**
+     * Getter/Setter
+     * @param {Boolean|Function} value whether the class should immediately subscribe to the `all` publication to feed its internal list
+     * @returns {Boolean}
+     */
+    feedNow( value ){
+        logger.verbose({ verbosity: AccountsManager.configure().verbosity, against: AccountsManager.C.Verbose.FUNCTIONS }, 'amOptions.feedNow()', arguments );
+        return this.base_gsBoolFn( 'feedNow', value, { default: amOptions._defaults.feedNow });
     }
 
     /**
@@ -109,6 +125,7 @@ export class amOptions extends AccountsCore.Options {
      */
     listActiveCheckboxes( value ){
         logger.verbose({ verbosity: AccountsManager.configure().verbosity, against: AccountsManager.C.Verbose.FUNCTIONS }, 'amOptions.listActiveCheckboxes()', arguments );
+        logger.warning( '\'listActiveCheckboxes\' instanciation option is deprecated starting with v2.5 in favor of setupTabular({ \'activeCheckboxes\' }). You should update your code.' );
         return this.base_gsBoolFn( 'listActiveCheckboxes', value, { default: amOptions._defaults.listActiveCheckboxes });
     }
 
@@ -119,6 +136,7 @@ export class amOptions extends AccountsCore.Options {
      */
     listFeedNow( value ){
         logger.verbose({ verbosity: AccountsManager.configure().verbosity, against: AccountsManager.C.Verbose.FUNCTIONS }, 'amOptions.listFeedNow()', arguments );
-        return this.base_gsBoolFn( 'listFeedNow', value, { default: amOptions._defaults.listFeedNow });
+        logger.warning( '\'listFeedNow\' instanciation option is deprecated starting with v2.5 in favor of \'feedNow\'. You should update your code.' );
+        return this.feedNow( value );
     }
 }